fix: 修复编译错误,使项目能够成功构建
- 修复字符串插值语法错误
- 修复 onChange 和 ForEach 的使用方式
- 添加缺失的 Combine 导入
- 修复 MenuBarManager 初始化问题
- 修复属性初始化顺序问题
- 添加缺失的 @objc 标记
- 修复字体设置 API 使用
- 修复 debounce 参数类型
- 修复 WindowGroup 初始化器歧义

+++ QuickNoteClone/ViewModels/NoteListViewModel.swift
@@ -116,7 +116,7 @@ class NoteListViewModel: ObservableObject {
 
     private func setupSearch() {
         searchCancellable = $searchText
-            .debounce(for: .milliseconds(Constants.Performance.maxSearchDelay * 1000), scheduler: RunLoop.main)
+            .debounce(for: .milliseconds(Int(Constants.Performance.maxSearchDelay * 1000)), scheduler: RunLoop.main)
             .sink { [weak self] _ in
                 self?.performSearch()
             }
